requirements.txt
annotated-types
anyio
click
fastapi
h11
idna
pydantic
pydantic_core
python-multipart
sniffio
starlette
typing-inspection
typing_extensions
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
requests
PyMuPDF
google-generativeai
langchain
langchain-community
faiss-cpu
sentence-transformers
huggingface-hub
pypdf
tf-keras